requirements.txt
mysql-connector-python>=8.0.0
configparser>=5.0.0